+++ db/__init__.py
@@ -1,12 +1,11 @@
-from sqlalchemy import create_engine
+from sqlalchemy import create_engine, event
+from sqlalchemy.engine import Engine
 from sqlalchemy.orm import sessionmaker
-from sqlalchemy.ext.declarative import declarative_base
 
 
 # engine = create_engine("sqlite+pysqlite:///:memory:", echo=True, future=True)
-engine = create_engine("sqlite:///axieST.db", echo=True, future=True)
+engine = create_engine("sqlite:///axieST.db", echo=False, future=True)
 Session = sessionmaker(bind=engine)
-Base = declarative_base()
 
 _session = None
 
@@ -19,13 +18,10 @@ def get_session():
     return _session
 
 
-# give to Base the ability to persist data on the db
-def save(self):
-    """Persist the current object to database
-    
-    """
-    session = get_session()
-    session.add(self)
-    session.commit()
-
-Base.save = save
+# enable foreign keys in SQLite3 - need to be set up to each connection
+# https://docs.sqlalchemy.org/en/13/dialects/sqlite.html#foreign-key-support
+@event.listens_for(Engine, "connect")
+def set_sqlite_pragma(dbapi_connection, connection_record):
+    cursor = dbapi_connection.cursor()
+    cursor.execute("PRAGMA foreign_keys=ON")
+    cursor.close()