+++ CHANGELOG.md
@@ -1,4 +1,7 @@
 Changelog
 
-# [next] unreleased
-- 
+# [0.1.0] 2022-01-04
+- added models: Scholar, Track
+- added command to manage input/output with Scholar CRUDL actions,
+  auto-gather of scholar performance (tracks), export of data (console, json
+  and csv)